Настраивает webpack, разбивает на модули
