+++ .gitignore
@@ -1,9 +1,16 @@
-* text eol=lf
+# Файлы и папки операционной системы
+.DS_Store
+Thumbs.db
 
-*.png binary
-*.jpg binary
-*.jpeg binary
-*.webp binary
-*.woff binary
-*.woff2 binary
+# Файлы редактора
+.idea
+*.sublime*
+.vscode
+
+# Вспомогательные файлы
+*.log*
 node_modules/
+
+# Собранные JS-файлы (бандл)
+public/bundle.js
+public/bundle.js.map
