+++ src/view/board.js
@@ -1,0 +1,11 @@
+const createBoardContainerTemplate = () => {
+  return (
+    `<section class="board container">
+
+    <div class="board__tasks"></div>
+    </section>`
+  );
+};
+
+export {createBoardContainerTemplate};
+
