+++ src/view/load-more.js
@@ -1,0 +1,7 @@
+const createLoadMoreTemplate = () => {
+  return (
+    `<button class="load-more" type="button">load more</button>`
+  );
+};
+
+export {createLoadMoreTemplate};
